Move Requirement to requirements.py
